+++ src/types.ts
@@ -15,3 +15,10 @@ export type PossibleArray<T> = T | T[];
 export type Opaque<K, T> = T & { __TYPE__: K };
 
 export type ISODateString = Opaque<'Date', string>;
+
+export type QualyNumericSeverity = 1 | 2 | 3 | 4 | 5 | undefined;
+
+/**
+ * Values that conform to the `Finding.numericSeverity` property.
+ */
+export type NormalizedNumericSeverity = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;